Extract base class from ArgumentConstraintOutsideCallSpecAnalyzer
